try new build script on real pipelines
